Fixed Face Behaviour Math calculations
Previously, only the target.orientation was used which was always 0. Now the correct angle between the agent's forward and agent->target vector is used.

+++ Source/GameAIProg/Movement/SteeringBehaviors/SteeringAgent.cpp
@@ -38,7 +38,7 @@ void ASteeringAgent::Tick(float DeltaTime)
 		if (!FMath::IsNearlyZero(output.AngularVelocity))
 		{
 			FRotator newRotation = GetActorRotation();
-			newRotation.Yaw += output.AngularVelocity * DeltaTime;
+			newRotation.Yaw += static_cast<double>(output.AngularVelocity) * DeltaTime;
 			SetActorRotation(newRotation);
 		}
 	}
